small refactoring 3
